+++ CHANGELOG.md
@@ -1,11 +1,15 @@
+## 3.8.2
+
+* SharedPreferences: Added get/set nullable string.
+
 ## 3.8.1
 
-* SharedPreference: Added get/set nullable int.
+* SharedPreferences: Added get/set nullable int.
 
 ## 3.8.0
 
 * Added LoggerAnalyticsService.
-* SharedPreference: Added get/set nullable bool.
+* SharedPreferences: Added get/set nullable bool.
 
 ## 3.7.3